requirements.txt
#aiohttp==3.6.2
#aiohttp-cors==0.7.0
#aiobotocore==0.9.4
#aioredis==1.3.1
#PyJWT==1.7.1
#redis==3.4.1
#envparse==0.2.0
#pytest==5.3.5
#pytest-cov==2.8.1
#python-magic==0.4.15
#Pillow==7.0.0
#pypng==0.0.20
#uvloop==0.14.0
#Werkzeug==1.0.0


aiohttp==3.4.4
aiohttp-cors==0.7.0
aiobotocore==0.9.4
aioredis==1.2.0
PyJWT==1.4.2
redis==2.10.5
raven-aiohttp==0.7.0
envparse==0.2.0
pytest==4.0.1
pytest-cov==2.6.0
python-magic==0.4.15
Pillow==7.0.0
pypng==0.0.18
uvloop==0.11.3
Werkzeug==0.14.1